+++ CHANGES.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3] - 2025-08-24
+
+### Changed
+
+- Better handling of alchemy ships
+- Handle HYP jumps in trade routes
+- Patched column sort into ship dialogue
+
 ## [1.2] - 2024-12-17
 
 ### Added